Update gpep-delete-passthrough-entry-after-submission.php
Fixed issue where fatal error was generated when token's entry did not exist.

+++ gp-easy-passthrough/gpep-delete-passthrough-entry-after-submission.php
@@ -8,6 +8,8 @@
 add_filter( 'gform_after_submission', function() {
 	if ( rgget( 'ep_token' ) && is_callable( 'gp_easy_passthrough' ) ) {
 		$entry = gp_easy_passthrough()->get_entry_for_token( rgget( 'ep_token' ) );
-		GFAPI::delete_entry( $entry['id'] );
+		if ( $entry ) {
+			GFAPI::delete_entry( $entry['id'] );
+		}
 	}
 } );
